Update TGDS homebrew on NTR / TWL hardware as of 2024:
-ToolchainGenericDS ecosystem is fully operational in NTR/TWL hardware, which means a lot of TWL codebase fixes were backported to NTR codebase.
-Enable GBA Macro setting for TGDS homebrew (NTR/TWL).
-Fixed a lot of segfaults on TGDS Filesystem POSIX operations.
-Reduce TGDS Filesystem POSIX filehandles to 5, to save memory.
-Reduce TGDS Filesystem POSIX filehandle names from 256 bytes to 128 bytes, to save memory.
-TGDS Filesystem POSIX filehandles are allocated statically, to prevent memory allocation issues.
-Default memory allocation calls are required to be initialized, otherwise an exception occurs.
-Fixed TGDS-multimediaplayer on NTR/TWL hardware
-Fixed WoopsiSDK + TGDS-multiboot v3 on NTR/TWL hardware (SDK: TGDS1.65)

Context:
https://bitbucket.org/Coto88/toolchaingenericds/issues/18/update-tgds-homebrew-in-ntr-mode-as-of

+++ arm9/source/project/src/Scene.h
@@ -35,11 +35,11 @@
 
 #define PI 3.14159
 
-#define OBJ_CRAB 0
-#define OBJ_OCTOPUS 1
-#define OBJ_STARFISH 2
-#define OBJ_FISH 3
-#define OBJ_PLANT 4
+#define OBJ_CRAB (int)(0)
+#define OBJ_OCTOPUS (int)(1)
+#define OBJ_STARFISH (int)(2)
+#define OBJ_FISH (int)(3)
+#define OBJ_PLANT (int)(4)
 
 struct Camera{
 	GLfloat verticalTilt;
@@ -114,6 +114,7 @@ extern void drawSphereCustom(float r, int lats, int longs);
 extern void drawCircle(GLfloat x, GLfloat y, GLfloat r, GLfloat BALL_RADIUS);
 extern void drawCylinder(int numMajor, int numMinor, float height, float radius);
 extern GLint DLSOLIDCUBE0_06F;
+
 extern void glut2SolidCube0_06f();
 extern int InitGL(int argc, char *argv[]); /// initialises OpenGL
 extern void setupTGDSProjectOpenGLDisplayLists();
